work on uploading image locally
Having some trouble setting up the image uploading on the local server. I have followed the steps found here: https://www.taniarascia.com/how-to-upload-files-to-a-server-with-plain-javascript-and-php/

but when running on localhost chrome, I get a 405 method not accepted error on POST,  but when I run it on Microsoft Edge, my request appears to go through. I think it may just be a Chrome issue, but I haven't found anything conclusive on it.  I'm not sure how to go about fixing this